RFE #24 support for multiple presentations within single repo.
A new parameter (p) on slideshow URLs accepts a relative path to a custom directory within a repo where PITCHME.* can be found for a given presentation. If (p) is not specified, the default behavior using the root directory within the repo is preserved.

This feature supports the creation and management of multiple presentations within a single repo (branch), potentially sharing common assets such as images, styles, etc.

The following changes were required to support this feature:

- Optional parameter (p) on slideshow urls
- Each (p) generates isolated print PDF
- Each (p) generates isolated offline ZIP
- Caching of YAML and Markdown is now (p) aware