Added support for code-block delimiter and code fragment highlights.
This commit add support for a new code-block delimiter that provides a simple way to inject any source code file found within the repo as a Markdown code-block on a presentation slide. The delimiter syntax looks as follows:

—?code=path/to/source/code.file

This commit also introduces support for code fragment highlighting. The CF-marker syntax, one or more of which must follow a code-block in your markdown, looks as follows:

@[fragment-range](optional-note)

Where fragment-range identifies a single line number, @[1], or a range of numbers, @[5-10]. The optional note is plain text that is displayed on the slide below the code-block when the fragment is in focus.